requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cosdata-client==0.2.2
cosdata-fastembed==0.7.1
cosdata-sdk==0.2.5
datasets==4.4.1
dill==0.4.0
dotenv==0.9.9
fastapi==0.122.0
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
loguru==0.7.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
numpy==2.3.5
onnxruntime==1.23.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
py_rust_stemmers==0.1.5
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
win32_setctime==1.2.0
xxhash==3.6.0
yarl==1.22.0
